+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Tweaked action buttons style in `struct` view to work better for touch devices
 - Changed `struct` view to open a link in a new tab
 - Added `limit` and `limitCollapsed` options support for `struct` view
+- Fixed pie chart in `signature` view popups to work in all browsers (#17)
 
 ## 1.0.0-beta.20 (09-09-2019)